+ Improved BlockJson#exists by "caching" textures that are being moved when making resource packs and checking block textures against that list as opposed to checking if the physical texture file exists. + No longer depending on ResourceLoader as resource managing have been moved to resourcepacks/NotEnoughBlocks and registering that folder as a FolderResourcePack + Added an option to set 'transform' in the blockstate templates. However this will be replaced immedately with making a new variant in the template file called 'item' and pointing that to a block model. Much easier and allows more options for rendering items in the inventory.
'Ported' some code from EnderCore to 1.8, however this will be removed in next push.

+++ src/main/java/com/kimfy/notenoughblocks/common/util/FileUtilities.java
@@ -87,9 +87,11 @@ public class FileUtilities
     }
 
     /**
-     * DOES NOT WORK FOR FILES PUT IN RESOURCE LOADER FOLDER
-     * Expected input: getFileAsString(new ResourceLocation("modid", "path/to/file.extension"));
+     * DOES NOT WORK FOR FILES PUT IN RESOURCE LOADER FOLDER - ONLY FOR RESOURCE PACKS AND SRC/MAIN/RESOURCES
      *
+     * Expected input: getFileAsString(new ResourceLocation("modid", "path/to/file.extension"));
+     * Base directory would be assets/modid/ - if your file is located in assets/modid/json/file.json
+     * you'd call getFileFromAssets(RL(modid, "json/file.json"));
      */
     public static String getFileFromAssets(ResourceLocation resourceLocation)
     {
@@ -141,6 +143,12 @@ public class FileUtilities
 
     public static void write(File file, String content)
     {
+        if (file.exists())
+        {
+            NotEnoughBlocks.logger.info("File: " + file.getName() + " already exists in: " + file.getPath());
+            return;
+        }
+
         try
         {
             FileWriter fileWriter = new FileWriter(file);
